Move EditTextExerciseViewTab
